If podman or docker is available and we are not in a container
We should re-exec this script inside a llama.cpp container (incomplete
in this commit), truth is these AI runtime evironment are complex to
set up by say installing a few packages, etc. The often require a
certain version of mesa, llama.cpp, kernel drivers, ROCm, the list goes
on and on depending on hardware. Although if one does want
to run outside a container and or say they are running on macOS we
should support that.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>