Handles stdin input
Be able to pass a prompt or a file by stdin like :
```
cat file | ramalama run model prompt
```

Works by prepending the input to the prompt as I didn't managed to make stdin
for llama.cpp work.
Prepending the input to the prompt is inspired by shell_gpt, that's the way they
do it.

Signed-off-by: Aurelien Rouene <aurelien@rouene.fr>