cli.py: remove errant slash preventing the loading of user conf file(s)
Hello, thanks for a great project.

I am not a Python coder, but the official Python docs say of os.path.join that "if a segment is an absolute path, then all previous segments are ignored and joining continues from the absolute path segment".  So, I expect "os.path.join(config_home, "/ramalama", "ramalama.conf")" to always elide the config_home (read: "~/.config") and fail to load user config files when XDG_CONFIG_HOME is not set.  I do not believe this is the intended behavior.  I suspect that removing the slash renders the code correct, but... I am not a Python coder.  Works for me, but needs review, please.

Best regards,
FNG

Signed-off-by: Fred N. Garvin, Esq. <184324400+FNGarvin@users.noreply.github.com>