Stop checking if command is running in container
Back when we were embedding the python code into the container image
we would check if we were already injected and not run a container,
the problem with this, is it is breaking when ramalama is run within a
toolbx container or running via podman or docker. If the user of
RamaLama wants to run without containers, then they should specify, and
the tool should not try to handle this on it's own.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2329826

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>