When using podman run add --pull=newer to update container image
The default for podman is to use --pull=missing, this means that if the
OCI Image was previously pulled, Podman will use it, even if there is a
newer image available.  Switching to "newer" means that podman will
compare the image on the OCI registry to the local one, and pull it
if it was updated.

Docker does not support this feature.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>