On ARM by default turn on GPU acceleration
On x86 we don't want to do this without a GPU because x86
integrated graphics have very limited access to VRAM and it's
normally not worth it. But ARM SoC's share memory between CPU and
GPU, meaning it's worth it generally. And we care mostly about
Apple Silicon where we want this on in podman machine.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>