We are displaying display driver info, scope creep
We have to be mindful of maintenance of the codebase. Display
drivers have nothing to do with AI acceleration. Don't show
display info such as "Color LCD" :

{
    "Engine": {
        "Name": null
    },
    "GPUs": {
        "Detected GPUs": [
            {
                "Cores": "18",
                "GPU": "Apple M3 Pro",
                "Metal": "Metal 3",
                "Vendor": "Apple (0x106b)"
            },
            {
                "GPU": "Color LCD"
            }
        ],
        "INFO": "No errors"
    },
    "Image": "quay.io/ramalama/ramalama",
    "Runtime": "llama.cpp",
    "Store": "/Users/ecurtin/.local/share/ramalama",
    "UseContainer": false,
    "Version": "0.0.19"
}

Not sure about the "macOS detection covers AMD GPUs" code. We could
have external GPUs potentially on macOS but even in that case the
code seems illogical.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>