Removed error wrapping in urlopen
By wrapping the raised error from urlopen() in an IOError, the
caller can not accurately handle that error anymore as its missing
the http response code for http errors, for example. Instead, we do
not handle any error at all in the HTTPClient for urlopen and let
the error bubble up. This way the caller such as the download_file()
can properly handle errors such as skipping retries for 404.

Signed-off-by: Michael Engel <mengel@redhat.com>