Try to choose a free port on serve if default one is not available
This change tries first to find if the default port 8080 is available.
If not, it tries to find an available free port in the range 8081-8090 in random order.
An error if no free port is found.

In case of success, the chosen port is printed out for the user.

This does not apply if the user chooses a port different from 8080.

Note that this check could be still not be enough if the chosen port is taken
by another process after our check, in that case we will still fail at a later phase as today.

Includes unit testing.

Closes #797

Signed-off-by: Andrea Decorte <adecorte@redhat.com>