Try to choose a random free serving port
This change tries to find an available free port in the range 8080-65536, and it does
10 tries in case the chosen port is taken, before returning an error if no free port is found.

In case of success, the chosen port is printed out for the user.

This does not apply if the user overrides the default port.

Note that this check could be still not be enough if the chosen port is taken
by another process after our check, in that case we will still fail at a later phase as today.

Includes unit testing.

Closes #797

Signed-off-by: Andrea Decorte <adecorte@redhat.com>