Try to choose a free serving port if the default one is not available
This change checks if the default port is already taken. If it's the case,
it tries to increment by 1 until finding a free port.
We also print out the chosen port for the user.

This does not apply if the user overrides the default port.

Note that this check could be still not be enough if the chosen port is taken
by another process after our check, in that case we will still fail at a later phase as today.

Includes unit testing.

Closes #797

Signed-off-by: Andrea Decorte <adecorte@redhat.com>