Add software to support using rag in RamaLama
This PR just installs the python requirements needed to play with the
rag_framework.py file.

I have not added the docling support yet, since that would swell the
size of the images.  Will add that in a separate PR.

Also remove pragmatic and begin conversion to new rag tooling.

Signed-off-by: Brian <bmahabir@bu.edu>
Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>