Added --chat-template-file support to ramalama serve
Relates to: https://github.com/containers/ramalama/issues/890
Relates to: https://github.com/containers/ramalama/issues/947

If a chat template file can be extracted from the gguf model or if specified by
the model repo, it will now be used in the ramalama serve command and mounted
into the container. It has been included in the generation of the quadlet and
kube files as well.

Signed-off-by: Michael Engel <mengel@redhat.com>