Fix handling of $RAMALAMA_CONTAINER_ENGINE
apple_vm has a side effect of setting the podman_machine_accel global
variable which is used when running and serving models. Currently if
the user sets RAMALAMA_CONTAINER_ENGINE to podman in an alternative path
the apple_vm code is not called, so the global variable is not set.

Fixes: https://github.com/containers/ramalama/issues/1040

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>