Docker running of containers is blowing up
Seems Docker does not support
--entrypoint=[] like podman.  --entrypoint "" seems to work on both
platforms.

When attempting to mimic --pull=newer on Docker we were pulling the
wrong image, we should be attempting to pull the accellerated image
not the default.

For some reason llama-run --threads X is blowing up in a docker
container with the option not being supported. This could be something
being masked inside of Docker containers that is not masked inside of podman
containers. Someone who understands what llama-run is doing with the
--threads option would need to look further into this.

This should fix the issue in CI that is blowing up Docker tests.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>