Remove files with colon in their name
The verify_checksum unit tests use files with a colon in their name. This
causes issues for Windows machines since file names/paths can not contain
this symbol. Therefore, these files have been removed and the tests create
these on the fly and only when not run on Windows machines.

Signed-off-by: Michael Engel <mengel@redhat.com>