Added .partial files to be listed for list command
Fixes: https://github.com/containers/ramalama/issues/1104

The model store iterates through all files in the ref files for
the list command. If a pull has been cancelled, then these point
to non-existent files even though there are might be already
partial files. To avoid this error, the model store list will
check for the files to exist and check for partial files.

Signed-off-by: Michael Engel <mengel@redhat.com>