Quote strings with spaces in debug mode
Currently if you run in Debug mode and attempt to cut and paste the
Podman or Docker line, the PROMPT field has a space with a > in it.
When pasted this causes issues since it is not properly quoted.

With this change the the command can be successfully cut and pasted.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>