Map url:// prefix of model name to URL class
Relates to: https://github.com/containers/ramalama/issues/1325

In the list models function only the url:// prefix is present.
Passing a listed model to the factory can not map this model
input correctly to the URL model class. Therefore, this gets
extended and the unit tests updated by appropriate cases.

Signed-off-by: Michael Engel <mengel@redhat.com>