Ignore errors when removing snapshot directory on failed creation
Relates to: https://github.com/containers/ramalama/issues/1508

remove_snapshot should never fail, therefore adding the ignore_errors=True.
Before removing a snapshot with ramalama rm an existence check is made. If
the model does not exist, an error will be raised to preserve the previous
behavior of that command.

Signed-off-by: Michael Engel <mengel@redhat.com>