Do not run with --tty when not in interactive mode
I have found that when running with nvidia the -t (--tty) option
in podman is covering up certain errors.  When we are not running
ramalama interactively, we do not need this flag set, and this
would make it easier to diagnose what is going on with users
systems.

Don't add -i unless necessary

Server should not need to be run with --interactive or --tty.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>