cli: honor --image parameter in "run" mode and with "--rag"
Previously, to account for "--rag" option and the "rag" mode, the
args.image was mangled at a late point. At that point, args.image does
not indicate whether the value was chosen by default (and needs
adjustment for "-rag") or whether it was explicitly specified by the
user. In the latter case, we would not want to overwrite the user's
choice.

In those cases, the parameter was not ignored entirely, but rather the
tag was stripped, "-rag" appended, and the default tag added.
Nonetheless, we should honor the original image. Granted, the user must
choose an image that is suitable (e.g. for "--rag" mode).

Similarly, the "run" mode entirely ignored the "--image" argument.

Now, image selection happens early while parsing command line arguments.
It is based on CONFIG, the cli args and the hardware detection in
common.py.

Signed-off-by: Thomas Haller <thaller@redhat.com>