install ramalama into containers from the current checkout
Copy the current checkout of the ramalama repo into the containers and use that for installation.
This removes the need for an extra checkout of the ramalama repo, and is consistent with the build
process used by container_build.sh (which used a bind-mount rather than a copy).

This keeps the version of ramalama in sync with the Containerfiles, and makes testing and CI more
useful.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>