Use model organization as namespace when pulling Ollama models
Relates to: https://github.com/containers/ramalama/issues/1278

Previously, the /library namespace was hard-coded into the Ollama API registry
head which caused pulling namespaced Ollama models to fail with a 404 error.
By mapping the model organization to the API calls namespace, this has been
resolved. In addition, the organization for Ollama models falls back to
"library" if no organization can be found in the model input.

Signed-off-by: Michael Engel <mengel@redhat.com>