konflux: push images to the quay.io/ramalama org after integration testing
The "bats-integration" pipeline is now run on every pull request and push to
the main branch. In the case of pushes, if the integration tests run successfully,
the images are pushed to a repo matching their component name in the quay.io/ramalama
org. Images will be tagged with the sha of the commit from which they were built.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>