Remove GGUF version check when parsing
In the GGUFParser the version of the GGUF model is being checked against
a hard-coded constant, currently v3. This prevents using older (or newer)
GGUF models. This check should not be done in the parser, but rather in
the used backend such as llama.cpp. If the backend supports it, its fine
and if not, then an error is being thrown anyway.

Signed-off-by: Michael Engel <mengel@redhat.com>