Fix assembling URLs and names for split models
Previously, the loop to assemble the split model files was
missing the first file of the list due to using the iteration
variable incorrectly. In addition, all files had been assigned
the same name, which led to overwriting the files.
In order to ensure that the urls and names are assembled correctly
unit tests have been added as well.

Signed-off-by: Michael Engel <mengel@redhat.com>