Explicit pull of base container before converted container build
Fhe oci container uses the quiet option in order to get the image
id properly. This, however, causes the convert command to seemingly
sleep for a very long time since it might pull the latest base
container image (which is quite large). In order to provide feedback
to the user, lets pull explicityl before the build.

Signed-off-by: Michael Engel <mengel@redhat.com>