fix parsing of CDI yaml
A previous commit removed pyyaml as a dependency and replaced it with code which parses
the yaml line-by-line looking for "name" and using the following token as the device name.
However, the code did not handle the hyphen prefix used in yaml lists, so it was ignoring
all entries in the yaml, preventing use of "cuda" acceleration. Fix the parser to correctly
handle hyphens.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>