Revert "Switch default ramalama image build to use VULKAN"
This change makes ramalama container generate garbage on MacOS, as was
originally reported at:

https://github.com/containers/ramalama/issues/1126

I believe the original promise was that VULKAN will work fine with
podman 5.5+, but I'm on 5.5.1 and still experience this problem.

DKOMPUTE_OPT_DISABLE_VULKAN_VERSION_CHECK is no longer supported by
upstream, so leaving it out of the revert.

This reverts commit b6d5e95e2c175311fc88cf18d337a5fd1727745b.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>