Fixed demo script
Relates to: https://github.com/containers/ramalama/pull/1876

Previously, no parameter was passed to the multi-model script function
so that the shellcheck threw an error due to the dirname $1 usage.
In order to fix this, the directory of the script is detected and
passed to assemble the firefox command with proper path.

Signed-off-by: Michael Engel <mengel@redhat.com>