Suppress stderr when chatting without container
llama.cpp is notorious of its stderr pollution, so it makes it nearly
impossible to use `--nocontainer run` with it.

I think it's ok to stderr when container is used because the output goes
to the container log, but not when running on the host where
llama-server shares the same tty.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>