Added the GGUF field tokenizer.chat_template for getting chat template
Relates to: https://github.com/containers/ramalama/issues/1882

The tokenizer.chat_template field in the GGUF metadata contains a jinja template.
This can be passed to llama.cpp for inference. In addition, this field is also
sometimes available in Ollama models so that a potentially erroneous conversion
is not needed.

Signed-off-by: Michael Engel <mengel@redhat.com>