Fix the application of codespell in "make validate".
codespell should not update files in place, it should only
spellcheck. Updating in place messes up the git tree and potentially
causes problems.

"mountns" is in both the ignore-words-list specified in the toml file
and in the .codespelldict. The latter has priority, which means that
codespell tries to correct the ignore-words-list in the toml
file. Assuming that the word belongs in the .codespelldict, it makes
no sense to list it in the ignore-words-list so remove it.  Also fix
the formatting of the list.

Signed-off-by: John Wiele <jwiele@redhat.com>