Do not set the ctx-size by default
llama.cpp is defaulting to ctx-size of 4096 and we were hard coding
2048, which means we were not using the default setting. This PR
changes to use ctx-size=0 which will not be specified in the command
unless the value is > 0, so llama-server default will be used.

Also we hard coded cache_reuse=256 with no way for user to override,
this PR adds support for cache_reuse being set in ramalama.conf and on
the command line.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>