rag: generate doc2rag command-line with CommandFactory, and execute it with Engine
Use the CommandFactory to generate the doc2rag command-line, and create a subclass of Engine
to handle executing it in a container.

Add a dedicated mapping of env vars to rag images in the config, and use that to select the
rag image in the cli, making image selection more consistent.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>