Added data-files path to default config dirs list
Relates to: https://github.com/containers/ramalama/issues/2026

Previousy, RamaLama did not use the configuration and inference spec
files installed in the pyproject.toml data-files directory. This led
to a failure for the inference spec when RamaLama was installed via
pip. Also, it didn't use the configuration file in there as well.

Signed-off-by: Michael Engel <mengel@redhat.com>