Prefer the embedded chat template for ollama models
If a chat template is embedded in gguf it is more likely to be compatible with
llama.cpp. Converting ollama go templates works for older models but the logic
has diverged significantly in more recent models with complex templates for
tool calling.

Also cleanup a regression in the needs_conversion logic from 0.13

Signed-off-by: Oliver Walsh <owalsh@redhat.com>