docker: fix list command for oci images when running in a non-UTC timezone
"docker images" outputs the CreatedAt field in the local timezone, which breaks the code in oci_tools.list_models() that
assumes the timezone is always UTC, as it is with podman. Explicitly set the TZ=UTC environment variable when running
"docker images" to avoid the error.

Also includes a minor formatting fix in list_manifests() to make debug output more readable.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>