docker: run doc2rag as the current user
"ramalama rag" runs "doc2rag" in a container, which writes the vector DB to a
temporary directory so it can be built into an image using "docker build".
Under docker the vector DB was owned by root, which made it unreadable by
"docker build". Pass the "--user" option so the files are owned by the current
user.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>