Add --port flag to ramalama run command
Fixes #2078

Users can now specify a custom port for the ramalama run command using
the --port or -p flag. This allows multiple models to run simultaneously
on the same machine without port conflicts.

The default port remains 8080, but users can override it with:
  ramalama run --port 8081 model_name

Updated documentation with the new option and example usage.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>