Added mapping of https to specific transport
Relates to: https://github.com/containers/ramalama/discussions/2104

In order to enable simply copying the huggingface and ollama URL
from the website and pull the repository, the scheme is detected
and then the prefix replaced by one of the specific transports.
This also detects if a specific file, e.g. .gguf, should be pulled
and doesn't perform the mapping in that case.

Signed-off-by: Michael Engel <mengel@redhat.com>