Fix --runtime cli arg parsing
The --runtime are needs to be set before processing the inference spec,
otherwise the llama.cpp inference spec is used regardless of the runtime
specified on the cli.

Split the cli parsing into two phases:
In the first pass only the global args are parsed and the CONFIG object is
updated (setting runtime etc...).
In the second pass all subcommands are parsed too, which will process the
inference spec.

Signed-off-by: Oliver Walsh <owalsh@redhat.com>