Fix RAG mount option incompatibility with Docker
Docker does not accept the mount option 'rw=true' in the --mount syntax,
resulting in the error:
  invalid argument "...rw=true" for "--mount" flag: unexpected key 'rw' in 'rw=true'

Since read-write is the default behavior for mounts in both Docker and
Podman, the 'rw=true' option can simply be omitted.

Changes:
- Removed 'rw=true' from type=bind mount in add_rag() for DB source
- Removed 'rw=true' from type=image mount in add_rag() for image source
- Updated comment to clarify default behavior

Update tests to remove rw=true mount option checks

Update system and e2e tests to match the removal of the rw=true
mount option in RAG functionality. Since read-write is the default
behavior for mounts, the tests no longer need to check for this
explicit option.

Changes:
- Updated test/system/070-rag.bats: Removed rw=true from mount checks
  on lines 62 and 79
- Updated test/system/040-serve.bats: Removed rw=true from mount check
  on line 547
- Updated test/e2e/test_serve.py: Removed rw=true from mount assertion
  on line 752

This PR was created with the help of Cursor AI.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>