konflux: on PRs, only build images when there is a change that's relevant to the image content
Avoid rebuilding all images on every PR by only triggering a build when there is a change that
will affect the content of the image. If an image is unaffected by a PR, the last successful
build triggered by a merge to the main branch will be used for testing.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>