[draft_model] Fix OCI mount options for draft_model
When supplying draft model the old code pretty much tried to bindmount
model with both src and destination in the end translating to same
location. Also there was no support for mulitple file models so when one
tried to use 440B as main model and 30B as draft model it would not be
possible.

To prevent file collisions for example for mmproj and chat_template we
use namespaced directory. As of now draft_models won't work with
multimodal models but we can still do speculative decoding when
disabling vision by passing --no-mmproj.

Signed-off-by: Lukas Bezdicka <lbezdick@redhat.com>