ci: fix podman-in-podman setup for nvidia GPUs
To run CI in Konflux we're using podman-in-podman to provide fine-grained control
over the test environment. Previously we were deleting the CDI hooks from the outer
podman environment because they were unnecessary.

Recent changes to the cuda image require the symlinks created by the hooks to be
present, so mount /usr/bin/nvidia-container-runtime and /usr/bin/nvidia-cdi-hook
into the outer podman environment and re-enable the hooks.

Signed-off-by: Mike Bonnet <mikeb@redhat.com>